+++ backend/canisters/gldt_swap/impl/src/updates/set_buy_back_canister.rs
@@ -4,8 +4,14 @@ use canister_tracing_macros::trace;
 pub use gldt_swap_api_canister::set_buy_back_canister::{
     Args as SetBuyBackCanisterArgs, Response as SetBuyBackCanisterResponse,
 };
-use ic_cdk::update;
+use ic_cdk::{query, update};
 use icrc_ledger_types::icrc1::account::Account;
+
+#[query(guard = "caller_is_authorized", hidden = true)]
+#[trace]
+async fn set_buy_back_canister_validate(args: SetBuyBackCanisterArgs) -> Result<String, String> {
+    serde_json::to_string_pretty(&args).map_err(|_| "invalid payload".to_string())
+}
 
 #[update(guard = "caller_is_authorized")]
 #[trace]