+++ backend/canisters/gldt_swap/CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### [unreleased]
 
+#### Change
+
+- Unhide set_buy_back_canister on candid
+
 ### [1.1.4] - 2025-03-11
 
 #### Add